move SearchParameters object into pattern repository
