+++ lib/search_results/models/models.dart
@@ -1,1 +1,1 @@
-export 'search_parameters.dart';
+
