+++ packages/pattern_repository/lib/src/core/config_map.dart
