+++ packages/pattern_repository/lib/src/core/core.dart
@@ -1,3 +1,4 @@
 export 'cartesian_product_iterable.dart';
 export 'compare.dart';
+export 'config_map.dart';
 export 'rotateable_list.dart';
