+++ packages/pattern_repository/lib/src/models/models.dart
@@ -1,5 +1,6 @@
 export 'pattern.dart';
 export 'pattern_constraint.dart';
 export 'prechac_direction.dart';
+export 'search_parameters.dart';
 export 'throw.dart';
 export 'throw_constraint.dart';
